refactor(#18): migrate authentication logic to useAuthStore
