+++ src/lib/bsky.ts
@@ -11,112 +11,6 @@ import type {
 } from '~/src/types/index';
 
 /**
- * Logs in a user to Bluesky using their credentials
- * @param identifier - Email address of the user
- * @param password - User's password
- * @returns {Promise<void>} - A Promise that resolves when login completes
- * @throws {Error} - Throws when rate limit is exceeded or other login errors occur
- */
-export async function loginUser(
-  identifier: string,
-  password: string
-): Promise<void> {
-  const authStore = useAuthStore();
-
-  authStore.setLoginError('');
-  const agent = AtpService.getAgent();
-
-  try {
-    if (!identifier.includes('@')) {
-      authStore.setLoginError(
-        'Please use your email address to login, not your handle'
-      );
-      return;
-    }
-
-    const { data: loginData, success } = await agent.login({
-      identifier,
-      password,
-    });
-
-    if (success) {
-      const { did: userDid, handle } = loginData;
-      authStore.setFormInfo(`Logged in as ${handle} with DID ${userDid}`);
-      authStore.setDid(userDid);
-      authStore.login();
-
-      localStorage.setItem('loginData', JSON.stringify({ loginData }));
-    } else {
-      authStore.setFormInfo('Login Failed');
-    }
-  } catch (error) {
-    if ((error as Error).message.includes('Rate Limit Exceeded')) {
-      authStore.setLoginError(
-        'Login failed: Rate limit exceeded. Please use your email address to login.'
-      );
-    } else {
-      authStore.setLoginError(`Login failed: ${(error as Error).message}`);
-    }
-    console.error('Login error:', error);
-  }
-}
-
-/**
- * Checks for an existing login session and restores it
- * @returns {Promise<void>} - No return value
- * @throws {Error} - If session data cannot be parsed or is invalid
- */
-export async function checkLoginSession(): Promise<void> {
-  const authStore = useAuthStore();
-
-  const storedData = localStorage.getItem('loginData');
-  if (!storedData) {
-    return;
-  }
-
-  try {
-    const parsedData = JSON.parse(storedData);
-
-    if (!parsedData || typeof parsedData !== 'object') {
-      throw new Error('Invalid session data format');
-    }
-
-    if (!parsedData.loginData) {
-      throw new Error('Missing login data in stored session');
-    }
-
-    const { loginData } = parsedData;
-
-    if (!loginData.did || !loginData.handle || !loginData.accessJwt) {
-      throw new Error('Incomplete login data in stored session');
-    }
-
-    authStore.setFormInfo(
-      `Logged in as ${loginData.handle} with DID ${loginData.did}`
-    );
-    authStore.setDid(loginData.did);
-    authStore.login();
-
-    // Set the auth token using the centralized API Service
-    AtpService.setAuthToken(loginData.accessJwt);
-
-    const jwtExpiry = getJwtExpiry(loginData.accessJwt);
-    if (jwtExpiry && isTokenExpiringSoon(jwtExpiry)) {
-      console.warn(
-        'Auth token expiring soon - user may need to re-authenticate'
-      );
-    }
-  } catch (error) {
-    console.error('Failed to restore session:', error);
-
-    localStorage.removeItem('loginData');
-
-    authStore.logout();
-    authStore.setFormInfo('Session could not be restored. Please login again.');
-  }
-}
-
-/**
  * Fetches the user's timeline from Bluesky
  * @returns {Promise<{displayData: DataObject, timelineJSON: string}>} - Object containing formatted timeline data and raw JSON
  * @throws {Error} - When user is not logged in, token has expired, or there's an error fetching the feed
@@ -169,7 +63,7 @@ export const getTimeline = async (): Promise<{
     };
   } catch (error) {
     if ((error as Error).message === 'Token has expired') {
-      handleSessionExpired();
+      authStore.handleSessionExpired();
     }
     console.error('Error fetching timeline:', error);
     throw new Error('Error fetching feed');
@@ -298,7 +192,7 @@ export const getLists = async (
   } catch (error) {
     listsStore.setIsFetching(false);
     if ((error as Error).message === 'Token has expired') {
-      handleSessionExpired();
+      authStore.handleSessionExpired();
     }
     console.error('Error fetching lists:', error);
     throw new Error('Error fetching lists');
@@ -435,7 +329,7 @@ export const getFollows = async (
   } catch (error) {
     followsStore.setIsFetching(false);
     if ((error as Error).message === 'Token has expired') {
-      handleSessionExpired();
+      authStore.handleSessionExpired();
     }
     console.error('Error fetching follows:', error);
     throw new Error('Error fetching follows');
@@ -602,18 +496,6 @@ const getCurrentListsPageData = (page: number, listsStore: ListsStore) => {
 };
 
 /**
- * Helper function to handle expired sessions
- * @returns {void} - No return value
- */
-const handleSessionExpired = (): void => {
-  const authStore = useAuthStore();
-
-  authStore.setFormInfo('Session expired. Please login again.');
-  localStorage.removeItem('loginData');
-  window.location.reload();
-};
-
-/**
  * Adds a user to a specified list
  * @param {string} userDid - The DID of the user to add
  * @param {string} listUri - The URI of the list to add the user to
@@ -659,7 +541,7 @@ export const addUserToList = async (
     return 'User successfully added to list';
   } catch (error) {
     if ((error as Error).message === 'Token has expired') {
-      handleSessionExpired();
+      authStore.handleSessionExpired();
     }
 
     console.error('Error adding user to list:', error);
@@ -747,7 +629,7 @@ export const addUsersToLists = async (
         });
       } catch (error) {
         if ((error as Error).message === 'Token has expired') {
-          handleSessionExpired();
+          authStore.handleSessionExpired();
         }
 
         results.push({
@@ -765,67 +647,3 @@ export const addUsersToLists = async (
 
   return results;
 };
-
-/**
- * Helper function to decode JWT and extract expiry time
- * @param {string} jwt - The JWT token
- * @returns {number|null} - Timestamp of token expiry or null if invalid
- */
-function getJwtExpiry(jwt: string): number | null {
-  try {
-    const base64Url = jwt.split('.')[1];
-    if (!base64Url) return null;
-
-    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-
-    const jsonPayload = decodeBase64(base64);
-
-    const payload = JSON.parse(jsonPayload);
-    return payload.exp ? payload.exp * 1000 : null;
-  } catch (error) {
-    console.error('Error decoding JWT:', error);
-    return null;
-  }
-}
-
-/**
- * Helper function to decode base64 safely
- * @param {string} str - The base64 string to decode
- * @returns {string} - The decoded string
- */
-function decodeBase64(str: string): string {
-  try {
-    if (typeof window !== 'undefined' && window.atob) {
-      return decodeURIComponent(
-        window
-          .atob(str)
-          .split('')
-          .map((c) => '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2))
-          .join('')
-      );
-    } else if (typeof Buffer !== 'undefined') {
-      return Buffer.from(str, 'base64').toString('utf-8');
-    } else {
-      const binaryString = atob(str);
-      return decodeURIComponent(
-        binaryString
-          .split('')
-          .map((c) => '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2))
-          .join('')
-      );
-    }
-  } catch (error) {
-    console.error('Error decoding base64:', error);
-    return '';
-  }
-}
-
-/**
- * Check if token is expiring within the next 5 minutes
- * @param {number} expiryTime - Token expiry timestamp
- * @returns {boolean} - True if token expires soon
- */
-function isTokenExpiringSoon(expiryTime: number): boolean {
-  const fiveMinutesInMs = 5 * 60 * 1000;
-  return Date.now() + fiveMinutesInMs > expiryTime;
-}
